Allow define one optional volume
For some cases and workloads we need to pass some modules to the task
definition. The volumes are passed as a list of maps to the
ecs_task_resource[1].

But the current syntax of terraform does not allow consume this value
directly from a variable. Trying to do so, we hit the issues described
in [2] and [3] (pending to report a specific bug).

But we found out that it works if we build the map structure directly
within the resource, and we use interpolation to consume the values of
the map.

Meanwhile the terraform project does not provide a definitive solution,
we will implement the following workaround:

 - The module will receive configuration for one unique module, being
   the default a empty map {}
 - If the map is empty, a dummy volume will be passed as `name=dummy`
   and `host_path=/tmp/dummy_volume`

This would cover our specific case in the short term, and can be later
easily adapted to use a optional list of volumes.

[1] https://www.terraform.io/docs/providers/aws/r/ecs_task_definition.html#volume
[2] https://github.com/hashicorp/terraform/issues/10407
[3] https://github.com/hashicorp/terraform/issues/7705#issuecomment-300564571